refactor: move project method inside aggregate context
