+++ dsl/src/commonMain/kotlin/it/unibo/collektive/field/Field.kt
@@ -1,8 +1,6 @@
 package it.unibo.collektive.field
 
 import it.unibo.collektive.ID
-import it.unibo.collektive.aggregate.AggregateContext
-import it.unibo.collektive.aggregate.ops.neighbouring
 
 /**
  * A field is a map of messages where the key is the [ID] of a node and [T] the associated value.
@@ -173,22 +171,4 @@ internal class SequenceBasedField<T>(
 
     override fun <R> mapOthersAsSequence(transform: (ID, T) -> R): Sequence<Pair<ID, R>> =
         others.map { (id, value) -> id to transform(id, value) }
-}
-
-/**
- * TODO.
- */
-fun <T> AggregateContext.project(field: Field<T>): Field<T> {
-    val others = neighbouring(0.toByte())
-    return when {
-        field.neighborsCount == others.neighborsCount -> field
-        field.neighborsCount > others.neighborsCount -> others.mapWithId { id, _ -> field[id] }
-        else -> error(
-            """
-                Collektive is in an inconsistent state, this is most likely a bug in the implementation.
-                Field ${field} with ${field.neighborsCount} neighbors has been projected into a context
-                with more neighbors, ${others.neighborsCount}: ${others.excludeSelf().keys}.
-                """.trimIndent().replace(Regex("'\\R"), " ")
-        )
-    }
 }
